add --std flag to autoharness
